[rename] 응답 dto 패키지명 responsedto로 수정 및 model 패키지 하위로 이동
