+++ src/main/java/com/zerobase/musinsamonitor/controller/ProductController.java
@@ -4,8 +4,8 @@ import com.zerobase.musinsamonitor.crawler.Crawler;
 import com.zerobase.musinsamonitor.crawler.MusinsaCrawler;
 import com.zerobase.musinsamonitor.crawler.constants.Category;
 import com.zerobase.musinsamonitor.crawler.dto.CrawledResult;
-import com.zerobase.musinsamonitor.dto.ProductAndPricesResponseDto;
-import com.zerobase.musinsamonitor.dto.ProductResponseDto;
+import com.zerobase.musinsamonitor.model.responsedto.ProductAndPricesResponseDto;
+import com.zerobase.musinsamonitor.model.responsedto.ProductResponseDto;
 import com.zerobase.musinsamonitor.service.CrawlingService;
 import com.zerobase.musinsamonitor.service.ProductService;
 import lombok.RequiredArgsConstructor;
