+++ src/main/java/com/zerobase/musinsamonitor/model/responsedto/ProductAndPricesResponseDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.musinsamonitor.dto;
+package com.zerobase.musinsamonitor.model.responsedto;
 
 import com.zerobase.musinsamonitor.repository.entity.Price;
 import com.zerobase.musinsamonitor.repository.entity.Product;
