+++ src/main/java/com/zerobase/musinsamonitor/model/responsedto/ProductResponseDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.musinsamonitor.dto;
+package com.zerobase.musinsamonitor.model.responsedto;
 
 import com.zerobase.musinsamonitor.repository.entity.Product;
 import java.io.Serializable;
